actor: move chat messages into message module
- Factored all chat-message-related types (`ChatMessage`, `ChatResponseChunk`, `ToolCallRequest`, `FunctionCallRequest`, etc.) into a new sub-module `message`.
- Added `mod message` and re-exported symbols with `pub use message::*`.
- Implemented `FromIterator<ChatResponseChunkDelta>`, `Default`, and `Extend` for `ChatMessage` to enable `collect`.
- Added a new test verifying stream deltas can be collected into a complete `ChatMessage`.
